fix(184): batch-4 — extract build_ast_weight_table, ast_cmd module, and integration tests
B3/B6: Extract `build_ast_weight_table()` into `ast_pipeline` library module.
The function encapsulates extract→stabilize→rekey→IDF→assemble so the pipeline
is reusable and testable in isolation. `cmd_ast_run` reduces to load→fetch→call→write.

B5: Add 5 integration tests in `ast_pipeline` covering: non-empty output from
real Rust fixture, multi-language corpus, empty corpus, high-threshold filtering,
and `generated_at` passthrough.

SF8: Extract AST subcommand handlers into `ast_cmd.rs`. `main.rs` goes from 562
to 399 lines, resolving the 500-line threshold warning.

SF1 (false positive): `stabilize()` clones ~100-300 short strings once in an
offline pass — negligible cost, no fix warranted.
