+++ crates/rskim-research/src/main.rs
@@ -12,14 +12,10 @@
 //! - `ast-codegen`: read ast_weights.json, generate ast_weights.rs for rskim-search
 //! - `ast-validate`: read ast_weights.json, run AST validation report
 
-use rskim_research::{
-    ast_codegen, ast_extract, ast_idf, ast_types, ast_validate, clone, codegen, config, extract,
-    idf, types, validate,
-};
+use rskim_research::{clone, codegen, config, extract, idf, types, validate};
 
 use serde::Serialize;
 
-use std::collections::HashMap;
 use std::path::PathBuf;
 
 use anyhow::Context;
@@ -143,14 +139,14 @@ fn main() -> anyhow::Result<()> {
             corpus_config,
             output,
             trigrams,
-        } => cmd_ast_run(corpus_dir, threshold, &corpus_config, output, trigrams),
+        } => ast_cmd::cmd_ast_run(corpus_dir, threshold, &corpus_config, output, trigrams),
 
         Commands::AstCodegen {
             json_path,
             workspace_root,
-        } => cmd_ast_codegen(json_path, workspace_root),
+        } => ast_cmd::cmd_ast_codegen(json_path, workspace_root),
 
-        Commands::AstValidate { json_path } => cmd_ast_validate(json_path),
+        Commands::AstValidate { json_path } => ast_cmd::cmd_ast_validate(json_path),
     }
 }
 
@@ -381,174 +377,15 @@ fn cmd_validate(json_path: Option<PathBuf>) -> anyhow::Result<()> {
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
-// AST subcommand handlers
+// AST subcommand handlers (separate module)
 // ─────────────────────────────────────────────────────────────────────────────
 
-fn cmd_ast_run(
-    corpus_dir: Option<PathBuf>,
-    threshold: f32,
-    corpus_config: &std::path::Path,
-    output: Option<PathBuf>,
-    collect_trigrams: bool,
-) -> anyhow::Result<()> {
-    let ast_config = config::load_ast_corpus_config(corpus_config)
-        .with_context(|| format!("loading AST corpus config from {}", corpus_config.display()))?;
-
-    let (corpus_dir_path, _temp_dir_guard) = resolve_corpus_dir(corpus_dir)?;
-
-    eprintln!(
-        "Cloning {} repos into {} ...",
-        ast_config.repos.len(),
-        corpus_dir_path.display()
-    );
-
-    let all_files = fetch_all_ast_files(&ast_config, &corpus_dir_path)?;
-
-    eprintln!(
-        "Loaded {} source files. Extracting AST n-grams...",
-        all_files.len()
-    );
-
-    let mut vocab = ast_types::NodeKindVocabulary::new();
-
-    let (raw_bigram_df_maps, raw_trigram_df_maps, corpus_stats) =
-        ast_extract::extract_ast_ngrams_from_corpus(&all_files, &mut vocab, collect_trigrams);
-
-    // Stabilize the vocabulary (sort alphabetically, reassign IDs) and get the
-    // old→new ID remap table. All bigram/trigram keys in the DF maps were encoded
-    // with pre-stabilize IDs and must be re-keyed before IDF computation.
-    let remap = vocab.stabilize();
-
-    eprintln!(
-        "Vocabulary: {} node kinds. Re-keying and computing IDF weights...",
-        vocab.len()
-    );
-
-    let mut bigram_weights_map: HashMap<String, Vec<ast_types::AstBigramWeight>> = HashMap::new();
-    let mut trigram_weights_map: HashMap<String, Vec<ast_types::AstTrigramWeight>> = HashMap::new();
-
-    let total_docs = corpus_stats.total_files;
-
-    for (lang, df_map) in &raw_bigram_df_maps {
-        let rekeyed = ast_types::rekey_bigram_df_map(df_map, &remap);
-        let weights = ast_idf::compute_ast_bigram_weights(&rekeyed, total_docs, threshold, &vocab);
-        eprintln!(
-            "  {lang}: {} bigrams (threshold={threshold})",
-            weights.len()
-        );
-        bigram_weights_map.insert(lang.clone(), weights);
-    }
-
-    for (lang, df_map) in &raw_trigram_df_maps {
-        let rekeyed = ast_types::rekey_trigram_df_map(df_map, &remap);
-        let weights = ast_idf::compute_ast_trigram_weights(&rekeyed, total_docs, threshold, &vocab);
-        eprintln!(
-            "  {lang}: {} trigrams (threshold={threshold})",
-            weights.len()
-        );
-        trigram_weights_map.insert(lang.clone(), weights);
-    }
-
-    let table = ast_types::AstWeightTable {
-        version: 1,
-        generated_at: chrono_now(),
-        vocabulary: vocab.kinds().to_vec(),
-        corpus_stats,
-        bigram_weights: bigram_weights_map,
-        trigram_weights: trigram_weights_map,
-    };
-
-    write_ast_weight_table(&table, output)?;
-    log_ast_summary(&table);
-
-    Ok(())
-}
-
-/// Print a compact per-language summary to stderr so the user can assess quality
-/// without running `ast-validate` separately. Mirrors the feedback provided by
-/// `cmd_run`'s `log_validation_summary` for the lexical pipeline.
+/// AST subcommand handlers: `ast-run`, `ast-codegen`, `ast-validate`.
 ///
-/// Reuses `run_ast_validation` for error-rate computation to avoid duplicating
-/// the formula that lives in `ast_validate`.
-fn log_ast_summary(table: &ast_types::AstWeightTable) {
-    let report = ast_validate::run_ast_validation(table);
-    eprintln!("=== AST weight table summary ===");
-    eprintln!("Vocabulary: {} node kinds", report.vocabulary_size);
-    for lang_report in &report.per_language {
-        let lang = &lang_report.bigram_distribution.language;
-        let bigrams = lang_report.bigram_distribution.count;
-        let trigrams = lang_report.trigram_distribution.count;
-        let error_rate = lang_report.error_node_rate * 100.0;
-        eprintln!("  {lang}: {bigrams} bigrams, {trigrams} trigrams, error_rate={error_rate:.2}%");
-    }
-}
-
-/// Clone/fetch all source files for the AST corpus using the AST extension walker.
-fn fetch_all_ast_files(
-    config: &config::CorpusConfig,
-    corpus_dir: &std::path::Path,
-) -> anyhow::Result<Vec<types::SourceFile>> {
-    let source = clone::AstGitCloneSource {
-        corpus_dir: corpus_dir.to_path_buf(),
-    };
-    fetch_files_parallel(config, &source)
-}
-
-/// Serialize the AST weight table to JSON and write to output path.
-fn write_ast_weight_table(
-    table: &ast_types::AstWeightTable,
-    output: Option<PathBuf>,
-) -> anyhow::Result<()> {
-    let output_path = output.unwrap_or_else(|| {
-        ast_codegen::default_ast_weights_json_path()
-            .unwrap_or_else(|_| PathBuf::from("ast_weights.json"))
-    });
-    write_json_table(table, output_path, "AST weight table")
-}
-
-fn cmd_ast_codegen(
-    json_path: Option<PathBuf>,
-    workspace_root: Option<PathBuf>,
-) -> anyhow::Result<()> {
-    let workspace_root = match workspace_root {
-        Some(p) => p,
-        None => codegen::find_workspace_root().context("auto-detecting workspace root")?,
-    };
-
-    let json_path = json_path
-        .unwrap_or_else(|| workspace_root.join("crates/rskim-search/data/ast_weights.json"));
-
-    let output_path = workspace_root.join("crates/rskim-search/src/ast_weights.rs");
-
-    eprintln!(
-        "Reading {} -> generating {}",
-        json_path.display(),
-        output_path.display()
-    );
-
-    ast_codegen::generate_ast_weights_rs(&json_path, &output_path)?;
-
-    eprintln!("Generated: {}", output_path.display());
-    Ok(())
-}
-
-fn cmd_ast_validate(json_path: Option<PathBuf>) -> anyhow::Result<()> {
-    let json_path = json_path.unwrap_or_else(|| {
-        ast_codegen::default_ast_weights_json_path()
-            .unwrap_or_else(|_| PathBuf::from("ast_weights.json"))
-    });
-
-    let raw = std::fs::read_to_string(&json_path)
-        .with_context(|| format!("reading {}", json_path.display()))?;
-
-    let table: ast_types::AstWeightTable =
-        serde_json::from_str(&raw).context("parsing ast_weights.json")?;
-
-    let report = ast_validate::run_ast_validation(&table);
-    ast_validate::print_ast_validation_report(&report);
-
-    Ok(())
-}
+/// Extracted into a separate file to keep `main.rs` below the 500-line
+/// threshold. Shared helpers (`resolve_corpus_dir`, `write_json_table`,
+/// `fetch_files_parallel`, `chrono_now`) are imported from `super::`.
+mod ast_cmd;
 
 /// Returns a UTC timestamp string for the `generated_at` field.
 fn chrono_now() -> String {
